requirements.txt
flask
pandas
openpyxl
python-docx
python-pptx
